task

+++ app/Http/Controllers/TasksController.php
@@ -15,12 +15,23 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+         $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('users.show', $data);
+        }else {
+            return view('welcome');
+        }
     }
+
+    
 
     /**
      * Show the form for creating a new resource.
@@ -117,8 +128,11 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::find($id);
+        $task = \App\Task::find($id);
+        
+        if (\Auth::id() === $task->user_id) {
         $task->delete();
+        }
         
         return redirect('/');
     }
